feat(camcontrol): Expose CAM bus
Ticket: #35998
